迁移清除缓存至系统内， 恢复`get_bot_status`方法
